+++ Branch-SDK/Branch-SDK/BNCEncodingUtils.m
@@ -258,7 +258,7 @@ static const short _base64DecodingTable[256] = {
 
     [encodedDictionary appendString:@"}"];
     
-    if ([BNCPreferenceHelper isDebug]) {
+    if ([[BNCPreferenceHelper preferenceHelper] isDebug]) {
         NSLog(@"encoded dictionary : %@", encodedDictionary);
     }
     
@@ -321,7 +321,7 @@ static const short _base64DecodingTable[256] = {
     [encodedArray deleteCharactersInRange:NSMakeRange([encodedArray length] - 1, 1)];
     [encodedArray appendString:@"]"];
     
-    if ([BNCPreferenceHelper isDebug]) {
+    if ([[BNCPreferenceHelper preferenceHelper] isDebug]) {
         NSLog(@"encoded array : %@", encodedArray);
     }