+++ Branch-SDK/Branch-SDK/BranchConnectDebugRequest.m
@@ -28,8 +28,10 @@
 }
 
 - (void)makeRequest:(BNCServerInterface *)serverInterface key:(NSString *)key callback:(BNCServerCallback)callback {
+    BNCPreferenceHelper *preferenceHelper = [BNCPreferenceHelper preferenceHelper];
+
     NSDictionary *params = @{
-        BRANCH_REQUEST_KEY_DEVICE_FINGERPRINT_ID: [BNCPreferenceHelper getDeviceFingerprintID],
+        BRANCH_REQUEST_KEY_DEVICE_FINGERPRINT_ID: preferenceHelper.deviceFingerprintID,
         BRANCH_REQUEST_KEY_DEVICE_NAME: [BNCSystemObserver getDeviceName],
         BRANCH_REQUEST_KEY_OS: [BNCSystemObserver getOS],
         BRANCH_REQUEST_KEY_OS_VERSION: [BNCSystemObserver getOSVersion],
@@ -37,7 +39,7 @@
         BRANCH_REQUEST_KEY_IS_SIMULATOR: @([BNCSystemObserver isSimulator])
     };
     
-    [serverInterface postRequest:params url:[BNCPreferenceHelper getAPIURL:BRANCH_REQUEST_ENDPOINT_CONNECT_DEBUG] key:key log:NO callback:callback];
+    [serverInterface postRequest:params url:[preferenceHelper getAPIURL:BRANCH_REQUEST_ENDPOINT_CONNECT_DEBUG] key:key log:NO callback:callback];
 }
 
 - (void)processResponse:(BNCServerResponse *)response error:(NSError *)error {
@@ -45,7 +47,7 @@
         NSLog(@"Failed to connect to debug: %@", error);
     }
     else {
-        [BNCPreferenceHelper setConnectedToRemoteDebug:YES];
+        [BNCPreferenceHelper preferenceHelper].isConnectedToRemoteDebug = YES;
 
         if (self.callback) {
             self.callback(YES, nil);