+++ Branch-SDK/Branch-SDK/BranchLoadActionsRequest.m
@@ -27,8 +27,9 @@
 }
 
 - (void)makeRequest:(BNCServerInterface *)serverInterface key:(NSString *)key callback:(BNCServerCallback)callback {
-    NSString *endpoint = [BRANCH_REQUEST_ENDPOINT_LOAD_ACTIONS stringByAppendingPathComponent:[BNCPreferenceHelper getIdentityID]];
-    [serverInterface getRequest:nil url:[BNCPreferenceHelper getAPIURL:endpoint] key:key callback:callback];
+    BNCPreferenceHelper *preferenceHelper = [BNCPreferenceHelper preferenceHelper];
+    NSString *endpoint = [BRANCH_REQUEST_ENDPOINT_LOAD_ACTIONS stringByAppendingPathComponent:preferenceHelper.identityID];
+    [serverInterface getRequest:nil url:[preferenceHelper getAPIURL:endpoint] key:key callback:callback];
 }
 
 - (void)processResponse:(BNCServerResponse *)response error:(NSError *)error {
@@ -45,12 +46,13 @@
         NSInteger total = [counts[BRANCH_RESPONSE_KEY_ACTION_COUNT_TOTAL] integerValue];
         NSInteger unique = [counts[BRANCH_RESPONSE_KEY_ACTION_COUNT_UNIQUE] integerValue];
         
-        if (total != [BNCPreferenceHelper getActionTotalCount:key] || unique != [BNCPreferenceHelper getActionUniqueCount:key]) {
+        BNCPreferenceHelper *preferenceHelper = [BNCPreferenceHelper preferenceHelper];
+        if (total != [preferenceHelper getActionTotalCount:key] || unique != [preferenceHelper getActionUniqueCount:key]) {
             hasUpdated = YES;
         }
         
-        [BNCPreferenceHelper setActionTotalCount:key withCount:total];
-        [BNCPreferenceHelper setActionUniqueCount:key withCount:unique];
+        [preferenceHelper setActionTotalCount:key withCount:total];
+        [preferenceHelper setActionUniqueCount:key withCount:unique];
     }
     
     if (self.callback) {