+++ Branch-SDK/Branch-SDK/BranchLogoutRequest.m
@@ -10,30 +10,54 @@
 #import "BNCPreferenceHelper.h"
 #import "BranchConstants.h"
 
+@interface BranchLogoutRequest ()
+
+@property (strong, nonatomic) callbackWithStatus callback;
+
+@end
+
 @implementation BranchLogoutRequest
 
+- (id)initWithCallback:(callbackWithStatus)callback {
+    if (self = [super init]) {
+        _callback = callback;
+    }
+    
+    return self;
+}
+
 - (void)makeRequest:(BNCServerInterface *)serverInterface key:(NSString *)key callback:(BNCServerCallback)callback {
+    BNCPreferenceHelper *preferenceHelper = [BNCPreferenceHelper preferenceHelper];
+
     NSDictionary *params = @{
-        BRANCH_REQUEST_KEY_DEVICE_FINGERPRINT_ID: [BNCPreferenceHelper getDeviceFingerprintID],
-        BRANCH_REQUEST_KEY_SESSION_ID: [BNCPreferenceHelper getSessionID],
-        BRANCH_REQUEST_KEY_BRANCH_IDENTITY: [BNCPreferenceHelper getIdentityID]
+        BRANCH_REQUEST_KEY_DEVICE_FINGERPRINT_ID: preferenceHelper.deviceFingerprintID,
+        BRANCH_REQUEST_KEY_SESSION_ID: preferenceHelper.sessionID,
+        BRANCH_REQUEST_KEY_BRANCH_IDENTITY: preferenceHelper.identityID
     };
 
-    [serverInterface postRequest:params url:[BNCPreferenceHelper getAPIURL:BRANCH_REQUEST_ENDPOINT_LOGOUT] key:key callback:callback];
+    [serverInterface postRequest:params url:[preferenceHelper getAPIURL:BRANCH_REQUEST_ENDPOINT_LOGOUT] key:key callback:callback];
 }
 
 - (void)processResponse:(BNCServerResponse *)response error:(NSError *)error {
     if (error) {
+        if (self.callback) {
+            self.callback(NO, error);
+        }
         return;
     }
 
-    [BNCPreferenceHelper setSessionID:response.data[BRANCH_RESPONSE_KEY_SESSION_ID]];
-    [BNCPreferenceHelper setIdentityID:response.data[BRANCH_RESPONSE_KEY_BRANCH_IDENTITY]];
-    [BNCPreferenceHelper setUserURL:response.data[BRANCH_RESPONSE_KEY_USER_URL]];
-    [BNCPreferenceHelper setUserIdentity:nil];
-    [BNCPreferenceHelper setInstallParams:nil];
-    [BNCPreferenceHelper setSessionParams:nil];
-    [BNCPreferenceHelper clearUserCreditsAndCounts];
+    BNCPreferenceHelper *preferenceHelper = [BNCPreferenceHelper preferenceHelper];
+    preferenceHelper.sessionID = response.data[BRANCH_RESPONSE_KEY_SESSION_ID];
+    preferenceHelper.identityID = response.data[BRANCH_RESPONSE_KEY_BRANCH_IDENTITY];
+    preferenceHelper.userUrl = response.data[BRANCH_RESPONSE_KEY_USER_URL];
+    preferenceHelper.userIdentity = nil;
+    preferenceHelper.installParams = nil;
+    preferenceHelper.sessionParams = nil;
+    [preferenceHelper clearUserCreditsAndCounts];
+    
+    if (self.callback) {
+        self.callback(YES, nil);
+    }
 }
 
 @end