+++ ChangeLog.md
@@ -1,5 +1,20 @@
 Branch iOS SDK change log
 
+- v0.10.0
+  * Adding an automatic deep linking feature, allowing devs to register a controller to be displayed based on keys in the open dictionary.
+  * Adding a delegate to the `BranchActivityItemProvider`, allowing devs to override link items based on selected sharing network.
+  * Fixed a potential crash w/ the persistence item if modified while saving.
+  * Deprecated some additional functions for the `getActivityItem:` methods, trying to move away from using "and" in method naming.
+  * Adding a check to prevent requests from being made when the SDK is in a bad state (missing device fingerprint or session).
+  * Exposed the `BNCConfig` header in the framework.
+
+- v0.9.3
+  * Clearing the Link Cache on logout; links shouldn't be shared between users.
+
+- v0.9.2
+  * Fixing check for isReferrable. No longer automatically setting to true for `handleDeepLink:`, and checking against `@0` since `nil` isn't possible.
+  * Making PreferenceHelper and non-singleton, and saving to file instead of using NSUserDefaults which made us prone to having our info wiped out from under us. Also keeping objects in memory, so that they don't need to be retrieved for each reference.
+
 - v0.9.1
   * Fixing an issue with archiving requests when requests are allocated too early.
   * Fixing a potential crash while calling close without a session.